refactor: promote OutputFormat to jhelm-core for REST/MCP reuse (#686)
The Helm-shaped output helper (snake_case release/history maps + json/yaml
mappers) lived in jhelm-cli, so only the CLI could produce `helm ... -o json`
shaped output. Move it to jhelm-core (org.alexmond.jhelm.core.output) so REST
and MCP can share the one implementation — the enabling step for #686 (Helm-
shaped output across all three surfaces).

Pure move: identical logic, new package; CLI command imports updated;
OutputFormatTest moved to core alongside it. No behavior change (only depends on
the core Release model + Jackson 3, both already on the core classpath).

Part of #686.
